lab-template/requirements.txt
jsonpickle
trio==0.14.0
typing_extensions
pyipv8==2.12
pyyaml
cryptography